Change the signatures of initial interpret and parsing fucntions.
Many of the functions required vm state, but in fact they should only care about
constant pool and entry point, that is the only thing they do when serializing.
VM state cares about state, and that is not present at the serialization time,
the vm.c functions should initialize this by themselves.