Add constructors.
The classes now have constructors. Only one constructor per class is
allowed. Constructor must always have a block (ie. single expression
results in an error). The block must either return nothing or None.

Internally, the constructor is called by a free function with the same
name as the class. This function creates the new instance and returns
it. The constructors themselves (the "init" methods) have as their
return value injected the "self" id. The proxy free function is also
done because constructor receive the "self".